FLASK - PyTesseract/requirements.txt
pytesseract
Flask
Pillow